docker_requirements.txt
pip
trimesh
open3d
kornia
pyrender
cython
scipy
git+https://github.com/mmatl/pyopengl.git
pybind11
tqdm
joblib
opencv-python
opencv-contrib-python
scikit-learn
freetype-py
pillow
numpy
pandas
matplotlib
pykdtree
torch
torchvision
tensorflow
tensorboard
tensorboardx
scikit-image
smplx==0.1.24
chumpy